nwatechfest-2023/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
openai==0.27.7
langchain==0.0.179
wolframalpha==5.0.0
google-search-results==2.4.2
wikipedia==1.4.0
httpx==0.24.1
gql==3.4.1
## The following requirements were added by pip freeze:
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
comm==0.1.3
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
frozenlist==1.3.3
graphql-core==3.2.3
h11==0.14.0
httpcore==0.17.2
idna==3.4
ipykernel==6.23.1
ipython==8.13.2
jaraco.context==4.3.0
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
more-itertools==9.1.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numexpr==2.8.4
numpy==1.24.3
openapi-schema-pydantic==1.2.4
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.8
Pygments==2.15.1
python-dateutil==2.8.2
PyYAML==6.0
pyzmq==25.0.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
stack-data==0.6.2
tenacity==8.2.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.6.1
urllib3==2.0.2
wcwidth==0.2.6
xmltodict==0.13.0
yarl==1.9.2
